requirements_comp.txt
# 語聲同行 競賽版 — 獨立依賴
# 建立日期: 2026-04-14
# 用途: 飛書 AI 校園挑戰賽（開放創新賽道）

# === ASR（從原專案複製，版本鎖定）===
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydub>=0.25.1

# Whisper + LoRA
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0

# === LLM 對話層（競賽新增）===
openai>=1.0.0          # 智譜 API 使用 OpenAI 兼容格式
zhipuai>=2.0.0         # 智譜官方 SDK（備選）

# === Web Demo 介面（競賽新增）===
gradio>=4.0.0

# === 工具 === #
python-dotenv>=1.0.0   # 環境變數管理
